// src/enums/worker.ts
export enum WorkerMessageType {
  PARSE_CSV = 'PARSE_CSV',
  PROCESS_SHAPES = 'PROCESS_SHAPES',
  CALCULATE_STOP_POSITIONS = 'CALCULATE_STOP_POSITIONS'
}

export enum WorkerResultType {
  PARSE_CSV_RESULT = 'PARSE_CSV_RESULT',
  PROCESS_SHAPES_RESULT = 'PROCESS_SHAPES_RESULT',
  CALCULATE_STOP_POSITIONS_RESULT = 'CALCULATE_STOP_POSITIONS_RESULT'
}